print_ray_v2.c to print_ray.c
